+++ R/styles.R
@@ -340,11 +340,6 @@ highlight <- function(x, i = NULL, j = NULL, color = "yellow", part = "body", so
 #'     ),
 #'     x
 #'   )
-#'
-#'   ft_2 <- flextable(x)
-#'   ft_2 <- color(ft_2, j = x$colname, color = scale)
-#'   ft_2 <- set_formatter_type(ft_2)
-#'   ft_2
 #' }
 color <- function(x, i = NULL, j = NULL, color, part = "body", source = j) {
   if (!inherits(x, "flextable")) {